natspec

+++ contracts/ZDAO.sol
@@ -6,10 +6,14 @@ import { GovernorSettings } from "@openzeppelin/contracts/governance/extensions/
 import { GovernorCountingSimple } from "@openzeppelin/contracts/governance/extensions/GovernorCountingSimple.sol";
 import { GovernorVotes, IVotes } from "@openzeppelin/contracts/governance/extensions/GovernorVotes.sol";
 import { GovernorTimelockControl, TimelockController } from "@openzeppelin/contracts/governance/extensions/GovernorTimelockControl.sol";
-import {GovernorVotesQuorumFraction } from "@openzeppelin/contracts/governance/extensions/GovernorVotesQuorumFraction.sol";
+import { GovernorVotesQuorumFraction } from "@openzeppelin/contracts/governance/extensions/GovernorVotesQuorumFraction.sol";
 
 /**
- * @title ZDAO
+ * @title ZDAO Governance Contract
+ * @notice A customizable governance contract based on OpenZeppelin's Governor contracts.
+ * @dev Extends OpenZeppelin's Governor contracts with various extensions for governance settings, voting, timelock control, and quorum fraction.
+ * See OpenZeppelin documentation: https://docs.openzeppelin.com/contracts/4.x/api/governance
+ * @custom:security-contact security@example.com
  */
 contract ZDAO is
     Governor,
@@ -19,14 +23,26 @@ contract ZDAO is
     GovernorTimelockControl,
     GovernorVotesQuorumFraction
 {
+    /**
+     * @notice Creates a new ZDAO governance contract.
+     * @param governorName The name of the governor instance.
+     * @param token The governance token that allows users to vote.
+     * @param timelock The timelock controller that handles proposal execution delay.
+     * @param delay_ The delay before voting starts (in blocks).
+     * @param votingPeriod_ The duration of the voting period (in blocks).
+     * @param proposalThreshold_ The minimum number of votes required to create a proposal.
+     * @param quorum_ The quorum fraction required for a proposal to pass.
+     * @dev Initializes the governor with settings and extensions.
+     * See OpenZeppelin Governor documentation: https://docs.openzeppelin.com/contracts/4.x/api/governance
+     */
     constructor(
         string memory governorName,
         IVotes token,
         TimelockController timelock,
-        uint delay_,
-        uint votingPeriod_,
-        uint proposalThreshold_,
-        uint quorum_
+        uint256 delay_,
+        uint256 votingPeriod_,
+        uint256 proposalThreshold_,
+        uint256 quorum_
     )
         Governor(governorName)
         GovernorVotes(token)
@@ -39,6 +55,12 @@ contract ZDAO is
         GovernorVotesQuorumFraction(quorum_)
     {}
 
+    /**
+     * @notice Returns the proposal threshold.
+     * @return The number of votes required to create a proposal.
+     * @dev Overrides the function from Governor and GovernorSettings.
+     * See OpenZeppelin GovernorSettings: https://docs.openzeppelin.com/contracts/4.x/api/governance#GovernorSettings
+     */
     function proposalThreshold()
         public
         view
@@ -48,6 +70,13 @@ contract ZDAO is
         return super.proposalThreshold();
     }
 
+    /**
+     * @notice Returns the current state of a proposal.
+     * @param proposalId The ID of the proposal.
+     * @return The ProposalState enum value.
+     * @dev Overrides the function from Governor and GovernorTimelockControl.
+     * See OpenZeppelin GovernorTimelockControl: https://docs.openzeppelin.com/contracts/4.x/api/governance#GovernorTimelockControl
+     */
     function state(uint256 proposalId)
         public
         view
@@ -57,6 +86,16 @@ contract ZDAO is
         return super.state(proposalId);
     }
 
+    /**
+     * @dev Executes a successful proposal.
+     * @param proposalId The ID of the proposal.
+     * @param targets The list of target addresses for calls.
+     * @param values The list of values (in wei) for calls.
+     * @param calldatas The list of calldata for calls.
+     * @param descriptionHash The hash of the proposal description.
+     * @dev Overrides the function from Governor and GovernorTimelockControl.
+     * See OpenZeppelin GovernorTimelockControl: https://docs.openzeppelin.com/contracts/4.x/api/governance#GovernorTimelockControl
+     */
     function _execute(
         uint256 proposalId,
         address[] memory targets,
@@ -76,6 +115,16 @@ contract ZDAO is
         );
     }
 
+    /**
+     * @dev Cancels a proposal.
+     * @param targets The list of target addresses for calls.
+     * @param values The list of values (in wei) for calls.
+     * @param calldatas The list of calldata for calls.
+     * @param descriptionHash The hash of the proposal description.
+     * @return The ID of the canceled proposal.
+     * @dev Overrides the function from Governor and GovernorTimelockControl.
+     * See OpenZeppelin GovernorTimelockControl: https://docs.openzeppelin.com/contracts/4.x/api/governance#GovernorTimelockControl
+     */
     function _cancel(
         address[] memory targets,
         uint256[] memory values,
@@ -89,6 +138,12 @@ contract ZDAO is
         return super._cancel(targets, values, calldatas, descriptionHash);
     }
 
+    /**
+     * @dev Returns the address of the executor.
+     * @return The address of the executor.
+     * @dev Overrides the function from Governor and GovernorTimelockControl.
+     * See OpenZeppelin GovernorTimelockControl: https://docs.openzeppelin.com/contracts/4.x/api/governance#GovernorTimelockControl
+     */
     function _executor()
         internal
         view
@@ -98,6 +153,13 @@ contract ZDAO is
         return super._executor();
     }
 
+    /**
+     * @notice Checks if a given interface is supported.
+     * @param interfaceId The interface identifier.
+     * @return True if the interface is supported, false otherwise.
+     * @dev Overrides the function from Governor and GovernorTimelockControl.
+     * See OpenZeppelin GovernorTimelockControl: https://docs.openzeppelin.com/contracts/4.x/api/governance#GovernorTimelockControl
+     */
     function supportsInterface(bytes4 interfaceId)
         public
         view
